fix trunk

+++ include/btrstd/string.h
@@ -72,7 +72,7 @@ void BTR_String_cropLeft(btr_string_s *, size_t count);
 void BTR_String_cropRight(btr_string_s *, size_t count);
 // Returns a pointer to a beginning of a UTF-8 character of the
 // given index. Returns `NULL` if the pointer is invalid.
-const char *BTR_String_charAt(btr_string_s *, int index);
+const char *BTR_String_charAt(btr_string_s *, long index);
 // Returns the amount of UTF-8 characters in the string.
 size_t BTR_String_len(btr_string_s *);
 // Tells if the function is empty.
